requirements.txt
python-dotenv
aiohttp
apscheduler
sanic
sanic-ext
asyncpg
hypercorn